sort out places where large byte[] buffers are used and long indexing needed, allow complex fastas for mono input, allow threadNum and preloadDb inputs

+++ MmseqsHelperLib/MmseqsIndexEntry.cs
@@ -1,3 +1,3 @@
 ﻿namespace MmseqsHelperLib;
 
-public record MmseqsIndexEntry(int Index, int StartOffset, int Length);
+public record MmseqsIndexEntry(int Index, long StartOffset, int Length);
